restructure into packages
