+++ .gitignore
@@ -4,6 +4,7 @@
 *.dll
 *.so
 *.dylib
+go6502
 
 # Test binary, built with `go test -c`
 *.test
