+++ emulators/apple1/apple1.go
@@ -1,0 +1,7 @@
+package apple1
+
+import "fmt"
+
+func Dummy() {
+	fmt.Println("apple1 here")
+}
