+++ go.mod
@@ -1,3 +1,3 @@
-module mos6502
+module github.com/KaiWalter/go6502
 
-go 1.17
+go 1.13
