+++ pkg/mos6502/load_test_rom.go
@@ -1,4 +1,4 @@
-package main
+package mos6502
 
 import (
 	"bufio"
